USST/physionet03/requirements.txt
numpy==1.16.2
scipy==1.2.1
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
torchvision